% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B1}
\alias{B1}
\title{Benchmark set B1}
\format{A vector containg 113 genes' names.}
\source{
The raw data can be downloaded from 
\url{http://www.cbs.dtu.dk/cellcycle/yeast_benchmark/benchmark.php}.
}
\usage{
data(B1)
}
\description{
List for yeast genes which are \bold{most} likely to be periodic 
(the benchmark set 1 in de Lichtenberg et al. (2005)).
}
\details{
A total of 113 genes previously identified as periodically
expressed in small-scale experiments. The set encompasses the
104 genes used by Spellman et al. (1998) and nine genes added
by Johansson et al. (2003).
}
\examples{
data(alpha)
data(B1)
alphaB1 <- alpha[rownames(alpha) \\\%in\\\% B1, ]

}
\references{
De Lichtenberg, U., Jensen, L. J., Fausboll, A., 
Jensen, T. S., Bork, P.,& Brunak, S. (2005). 
Comparison of computational methods for the identification 
of cell cycle-regulated genes. Bioinformatics, 21(7), 1164-1171.
}
\keyword{datasets}
