% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHRegLmrob.R
\name{fitHRegLmrob}
\alias{fitHRegLmrob}
\title{Robust Fitting of Four Parameter Harmonic Regression}
\usage{
fitHRegLmrob(z, t = 1:n, K = max(500, floor(length(z)/2)), ncpu = 1,
  lambdaRange = c(1/n, 0.45))
}
\arguments{
\item{z}{series.}

\item{t}{Time points.}

\item{K}{number of subintervals.}

\item{ncpu}{number of compute nodes for use with parallel.}

\item{lambdaRange}{range of frequencies inside (0,0.5), see Details.}
}
\value{
list with the following components
\itemize{
\item coefficients const, A, B, lambda
\item RSq
\item residuals
\item periodogram
\item z
\item t
\item K
\item model
\item lambdaRange
}
}
\description{
Estimates mu, A, B and lambda in the harmonic regression,
\eqn{y(t)=mu+A*cos(2*pi*lambda*t)+B*sin(2*pi*lambda*t)+e(t)},
where e(t) is assumed IID mean zero and constant variance.
The \code{robustbase::lmrob} function is used to fit the three parameter 
harmonic regression of a frequency grid. The optimal frequency corresponds 
to the harmonic regression with the smallest deviance.
The accuracy is controlled K, usually K=500, is sufficient.
}
\details{
The three parameter harmonic regression model is fitted 
for each frequency in the grid which is a subset of the 
equi-spaced point \eqn{{j/(2K+1), j=1,\dots,K}} possibly excluding low and
high frequencies as specified by the lambdaRange argument. Frequencies lower
than the first Fourier frequency, 1/n, where n is the length time series,
n=length(z) are usually due to trend while high frequencies are often not of
interest and the numerical algorithm for fitting the 3-parameter harmonic
regression is sometimes inaccurate at high. 
The function \link[robustbase]{lmrob} uses the MM robust method
which is also implemented in \link[MASS]{rlm}. 
The algorithm implemented \link[robustbase]{lmrob} provides 
efficient computation. The MM robust
regression method is robust as well as resistant to very large outliers.
}
\examples{
z<-c(0.42, 0.89, 1.44, 1.98, 2.21, 2.04, 0.82, 0.62, 0.56, 0.8, 1.33)
fitHRegLmrob(z)
#
#on multicore pcs, more the package parallel may be used for the
#grid computation but unless n is very large this is not recommended.
\dontrun{ #adjust ncpu
system.time(ans1 <- fitHRegLmrob(z)) #2.89 sec on my computer
system.time(ans2 <- fitHRegLmrob(z, ncpu=8)) #2.65 sec 
identical(ans1, ans2)
}
}
\references{
Venables, W. N. and Ripley, B. D. (2002) 
Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
\link[robustbase]{lmrob}, \link[MASS]{rlm}
}
\author{
Yuanhao Lai and A.I. McLeod
}
\keyword{ts}
