% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHRegLS.R
\name{fitHRegLS}
\alias{fitHRegLS}
\title{LS Fit for Four Parameter Harmonic Regression}
\usage{
fitHRegLS(z, t = 1:n, K = max(500, floor(length(z)/2)), ncpu = 1,
  lambdaRange = c(1/n, 0.45), exactQ = FALSE)
}
\arguments{
\item{z}{series.}

\item{t}{Time points.}

\item{K}{number of subintervals.}

\item{ncpu}{number of compute nodes for use with parallel.}

\item{lambdaRange}{range of frequencies inside (0,0.5), see Details.}

\item{exactQ, }{default setting is FALSE.
It indicates whether nolinear optimizer is used to 
obtain the final estimate of the frequency.}
}
\value{
list with the following components
\itemize{
\item coefficients const, A, B, lambda
\item residuals
\item periodogram
\item z
\item t
\item K
\item model
\item lambdaRange
}
}
\description{
Estimates mu, A, B and lambda in the harmonic regression,
\eqn{y(t)=mu+A*cos(2*pi*lambda*t)+B*sin(2*pi*lambda*t)+e(t)},
where e(t) is assumed NID mean zero and constant variance. 
The default algorithm is enumerative.
}
\details{
The QR decomposition is used to efficiently compute the residual
sum of squares on the grid. When \code{exactQ=TRUE}, a NLS
algorithm is used and NLS is indicated in the output title.
}
\examples{
z<-c(0.42, 0.89, 1.44, 1.98, 2.21, 2.04, 0.82, 0.62, 0.56, 0.8, 1.33)
fitHRegLS(z)
#
#on multicore pcs, more the package parallel may be used for the
#grid computation but unless n is very large this is not recommended.
\dontrun{ #adjust ncpu
system.time(ans1 <- fitHRegLS(z)) #0.06 sec on my computer
system.time(ans2 <- fitHRegLS(z, ncpu=8)) #1.67 sec 
identical(ans1, ans2)
}

}
\seealso{
\code{\link[base]{qr}}
}
\author{
Yuanhao Lai and A.I. McLeod
}
\keyword{ts}
