% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitFisherG.R
\name{GetFitFisherG}
\alias{GetFitFisherG}
\title{Fisher's g test}
\usage{
GetFitFisherG(z)
}
\arguments{
\item{z}{Input series, no missing values.}
}
\value{
Vector of length 2 containing the g test statistic
and the frequency for the maximum periodogram.
}
\description{
Calculates p-value using Fisher's g-test for a single frequency.
}
\details{
Program is interfaced to C programme (\code{\link{fft}}) for efficient computation.
}
\references{
Fisher, R.A. (1929). Tests of significance in harmonic analysis. 
Proc. Roy. Soc. A, 125, 54-59.
}
\author{
A.I. McLeod
}
\keyword{internal}
