% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitRobustG.R
\name{GetFitRobustG}
\alias{GetFitRobustG}
\title{Compute the Robust g test statistic}
\usage{
GetFitRobustG(y)
}
\arguments{
\item{y}{Vector containing the series.}

\item{freqSet}{The set to search frequencies.}
}
\value{
Vector of length 2 containing the Robust g test statistic 
and the frequency for the maximum periodogram.
}
\description{
The Robust g test statistic is computed for testing for periodicity.
}
\details{
This is the a modified version of the function robust.g.test() from the GeneCycle package.
}
\references{
Ahdesmaki, M., Lahdesmaki, H., Pearson, R., Huttunen, H., 
and Yli-Harja O.(2005). 
 \emph{BMC Bioinformatics} \bold{6}:117. 
 \url{http://www.biomedcentral.com/1471-2105/6/117}
}
\author{
Yuanhao Lai
}
\keyword{internal}
