% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MStats.R
\name{MStats}
\alias{MStats}
\title{Compute the test statistic based on the robust regression}
\usage{
MStats(z, lambdaRange = seq(1/N, 0.5 - 1/N, length.out = 2 * N))
}
\arguments{
\item{z}{A vector or a matrix containg series as columns.}

\item{lambdaRange}{A vector containing candidate frequencies.}
}
\value{
matrix with two columns, where the first column contains test statistics
and the second column contains estimated frequencies.
}
\description{
test statistic based on rank
}
\details{
TBA
}
\references{
TBA
}
\author{
Yuanhao Lai
}
\keyword{internal}
