% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hreg.R
\name{plot.hreg}
\alias{plot.hreg}
\title{Plot Diagnostics for "hreg" Object}
\usage{
\method{plot}{hreg}(x, type = c("fit", "pg", "acf", "nplot"), ...)
}
\arguments{
\item{x}{S3 class "hreg" object.}

\item{type}{Type of diagnostic plot - see details.}

\item{...}{Other parameters to be passed through to predicting functions.}
}
\value{
No value is returned. Plots are produced as side-effect.
}
\description{
Four plots of dignostic checks are available to check the adequacy of the
fitted four parameter harmonic regression model.
}
\details{
The diagnostic plot types are: "fit", "pg", "acf", "nplot", "cp", "rf" corresponding
to fitted model and data, periodogram, autocorrelation, normal probability plot, 
cumulative periodogram test, residual-fit spread plot.
}
\examples{
ans <- hreg(nottem)
plot(ans)
                                                       
}
\seealso{
\code{\link{hreg}}.
}
\author{
A.I. McLeod and Yuanhao Lai.
}
\keyword{ts}
