% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semihregScam.R
\name{predict.semihregScam}
\alias{predict.semihregScam}
\title{Prediction of the semiparametric harmonic regression with scam}
\usage{
\method{predict}{semihregScam}(object, newt, ...)
}
\arguments{
\item{object}{An object of class "semihregScam".}

\item{newt}{A vector of sampled time points for predicted values.}

\item{...}{Other parameters to be passed through to predicting functions.}
}
\value{
A matrix of two column, where
the first column contains the sample time points and
the second column contains the corresponding predicted values.
}
\description{
Return predicted values of the
semiparametric harmonic regression with scam,
The prediction procedure utilizes the function \code{predict.scam}
in the package \code{scam}.
}
\examples{
# Simulate the series
set.seed(193)
g <- function(x){-2*exp(-x^3)}
n <- 50
t <- 1:n
lambda0 <- 0.123
phi0 <- 0.2*2*pi
u <- cos(2*pi*lambda0*t+phi0)
e <- rnorm(n)
y <- g(u)+e

# Fit the semi-harmonic regression
fit1 <- semihregScam(y,t,lambda0=lambda0,bs="mpi",family=gaussian())
names(fit1)
fit1$esty

# Prediction for the continuous time.
newt <- seq(1,n,0.05) 
head( predy <- predict(fit1,newt)) 

plot(newt,g(cos(2*pi*lambda0*newt+phi0)),
     type="b",pch=19,main="Semiparametric with scam",
     xlab="t", ylab="g(u)",
     ylim=c(min(predy[,2])-1,max(predy[,2])+0.5) )
lines(newt,predy[,2],col="red",type="b")

}
\seealso{
\code{\link{semihregScam}}, \code{\link[scam]{scam}}.
}
\author{
Yuanhao Lai
}
