% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueRSR.R
\name{predictRSR}
\alias{predictRSR}
\title{Quantile computation for the RSR model}
\usage{
predictRSR(n, tableRSR)
}
\arguments{
\item{n}{Length of series.}

\item{tableRSR}{The RSR fitted models for the required statistic.}
}
\value{
a vector of all the estimated quantiles at pc for 
        specidifed sample sizes n from a given RSR model.
}
\description{
Compute the estimated quantiles of the required statistics from a given RSR model
}
\author{
Yuanhao Lai
}
\keyword{internal}
