% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hreg.R
\name{print.hreg}
\alias{print.hreg}
\title{Print Method for "hreg" Object}
\usage{
\method{print}{hreg}(x, ...)
}
\arguments{
\item{x}{S3 class "hreg" object.}

\item{...}{Other parameters to be passed through to printing functions.}
}
\value{
A terse summary is displayed.
}
\description{
Print estimates of the fitted four pararmeter harmonic regression model.
}
\examples{
ans <- hreg(nottem)
ans     
                                                 
}
\seealso{
\code{\link{hreg}}.
}
\author{
A.I. McLeod and Yuanhao Lai.
}
\keyword{ts}
