% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueRSR.R
\name{pvalueRSR}
\alias{pvalueRSR}
\title{P-value computation for the RSR model}
\usage{
pvalueRSR(n, Statistic, tableRSR)
}
\arguments{
\item{n}{Length of series.}

\item{Statistic}{A list of observed statistics.}

\item{tableRSR}{The RSR fitted models for the required statistic.}
}
\value{
A list of p-values.
}
\description{
Compute the estimated p-value of the observed statistics from a given RSR model
}
\author{
Yuanhao Lai
}
\keyword{internal}
