% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitRobustG.R
\name{rankSpectrum}
\alias{rankSpectrum}
\title{The robust rank-based spectral}
\usage{
rankSpectrum(x)
}
\arguments{
\item{x}{Series.}

\item{freqSet}{The set to search frequencies.}
}
\value{
A numerical value of the robust rank-based spectral.
}
\description{
This function implements the robust, rank-based spectral.
}
\keyword{internal}
