% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPhaseToAB.R
\name{RPhaseToAB}
\alias{RPhaseToAB}
\title{Convert harmonic regression parameters to R*cos(2*pi*f*t+phi) form}
\usage{
RPhaseToAB(RPhi)
}
\arguments{
\item{RPhi}{Vector containing the R and phi coefficients respectively.}
}
\value{
vector c(A, B)
}
\description{
The harmonic regression is usually fitted as 
   \eqn{z(t) = mu + A*cos(2*pi*f*t) + B*sin(2*pi*f*t)}
but is more meaningfully expressed as
\eqn{z(t) = mu + R*cos(2*pi*f*t + phi)},
where R and phi are the amplitude and phase.
This function converts the parameters in the R-phase form to the AB form.
}
\examples{
#check by back-transform
harper:::RPhaseToAB(harper:::ABToRPhase(c(-4.205,1.075)))
                                                        
}
\author{
Yuanhao Lai and A. I. McLeod
}
\keyword{internal}
