% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semihregScair.R
\name{semihregScair}
\alias{semihregScair}
\title{Fit the semiparametric harmonic regression with monotone index model}
\usage{
semihregScair(y, t, lambda0, shape = c("l", "in", "de", "cvxin", "cvxde",
  "ccvin", "ccvde"), family = gaussian(), ...)
}
\arguments{
\item{y}{A vector containg series as columns.}

\item{t}{A vector of sampled time points.}

\item{lambda0}{The frequency of the harmonic compoenent.}

\item{shape}{The shape constraint for the function g. 
See \code{\link[scar]{scair}} in the R package \code{scar}.}

\item{family}{a description of the error distribution and link function to be used in the model.
Default is gaussian(). See \code{\link[scar]{scair}}.}

\item{...}{Additional parameters passed to \code{\link[scar]{scair}}.}
}
\value{
Object of class "semihregScair" produced.

An object of class "semihregScair" is a list 
containing the following components:

\item{harScair}{ An object of class \code{scair}. 
See \code{\link[scar]{scair}} in the R package \code{scar}.}
\item{lambda0}{The frequency of the harmonic compoenent.}
\item{phi0}{The estimated phase/2pi of the harmonic compoenent.}
\item{esty}{estimated mean for the observations.}
}
\description{
Fit the semiparametric harmonic regression with monotone 
generalzied additive index model by maximizing non-parametric likelihood.
It utilizes the function \code{\link[scar]{scair}} in the R package \code{scar}.
}
\details{
Given frequency \eqn{\lambda},
fit a monotonic single index model with the likelihood method, i.e.,
\deqn{E[y_t] = \mu+g( A\cos(2\pi\lambda t)+B\sin(2\pi \lambda t) ),\ t=1,...,n,}
where \eqn{y_{t}} consists of independet random variables, 
\eqn{a_{1}^2+a_{2}^2=1},
\eqn{g(\cdot)} is an unknown strictly monotone function in the range [0,1], 
\eqn{0<f<0.5}, \eqn{A^2 + B^2 =1} and \eqn{\sigma>0}.

This is done by using the R package \code{scar}.
For more details, see \code{\link{scair}} 
in the R package \code{scar}.
}
\examples{
# Simulate the series
set.seed(193)
g <- function(x){-2*exp(-x^3)}
n <- 50
t <- 1:n
lambda0 <- 0.123
phi0 <- 0.2*2*pi
u <- cos(2*pi*lambda0*t+phi0)
e <- rnorm(n)
y <- g(u)+e

# Plot ther series
par(mfrow=c(1,2))
plot(t,g(u),type="b")
plot(t,y,type="b")
par(mfrow=c(1,1))

# Pre-estimate the frequency and the phase
lambdaRange <- seq(1/n,0.5-1/n,length.out = 5*n)
paraRankLS <- GetFitRankLS(y,t=1:n,lambdaRange)
paraRankLS    

# Fit the semi-harmonic regression
fit1 <- semihregScair(y,t,lambda0=paraRankLS[2],shape="in",iter=100)
names(fit1)
fit1$esty

# Prediction for the continuous time.
newt <- seq(1,n,0.05) 
head( predy <- predict(fit1,newt)) 

plot(newt,g(cos(2*pi*lambda0*newt+phi0)),
     type="b",pch=19,main="Semiparametric with scair",
     xlab="t", ylab="g(u)",
     ylim=c(min(predy[,2])-1,max(predy[,2])+0.5) )
lines(newt,predy[,2],col="red",type="b")


}
\references{
Chen, Y. and Samworth, R. J. (2014). Generalised additive and index models with shape constraints. arXiv:1404.2957.
}
\seealso{
\code{\link[scar]{scair}}.
}
\author{
Yuanhao Lai
}
\keyword{ts}
