% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semihregScam.R
\name{semihregScam}
\alias{semihregScam}
\title{Fit the semiparametric harmonic regression with monotone 
generalzied additive model}
\usage{
semihregScam(y, t, lambda0, k = -1, m = 2, bs = c("mpi", "mpd"),
  family = gaussian())
}
\arguments{
\item{y}{A vector containg series as columns.}

\item{t}{A vector of sampled time points.}

\item{lambda0}{The frequency of the harmonic compoenent.}

\item{k}{the dimension of the basis used to represent the smooth term.
See \code{\link[scam]{smooth.construct.mpi.smooth.spec}} in the R package \code{scam}.}

\item{m}{m+1 is the order of the B-spline basis.
See \code{\link[scam]{smooth.construct.mpi.smooth.spec}} in the R package \code{scam}.}

\item{bs}{specify the usage of monotone increasing/decreasing P-splines ("mpi"/"mpd") 
 as the penalized smoothing basis.
See \code{\link[scam]{shape.constrained.smooth.terms}} in the R package \code{scam}.}

\item{family}{A family object specifying the distribution of \code{z} and link to use in fitting. 
See \code{\link[stats]{glm}} and \code{\link[stats]{family}} for more details.}
}
\value{
Object of class "semihregScam" produced.

An object of class "semihregScam" is a list 
containing the following components:

\item{harScam}{ An object of class \code{scam}. 
See \code{\link[scam]{scam}} in the R package \code{scam}.}
\item{lambda0}{The frequency of the harmonic compoenent.}
\item{phi0}{The estimated phase/2pi of the harmonic compoenent.}
\item{esty}{estimated mean for the observations.}
}
\description{
Fit the semiparametric harmonic regression with monotone 
generalzied additive model.
It utilizes the R package \code{scam}.
The estimation procedure iterates between the coefficient of the index term 
and the smooth monotone P-spline function.
}
\details{
Given frequency \eqn{\lambda} and phase \eqn{\phi},
fit a monotonic smoothing term of sinuosoid with 
a monotone increasing/decreasing P-splines, i.e.,
\deqn{E[y_t] = g(a_{1} cos(2\pi\lambda t)+a_{2} sin(2\pi\lambda t) ),\ t=1,...,n,}
where \eqn{y_{t}} consists of independet random variables, 
\eqn{a_{1}^2+a_{2}^2=1},
\eqn{g(\cdot)} is an unknown strictly monotone function in the range [-1,1], 
\eqn{0<f<0.5}, \eqn{\phi\in [0,2\pi]} and \eqn{\sigma>0}.
This can be done by using the R package \code{scam}
in the R package \code{scam}.
Since the frequency \eqn{\lambda} is prespecified in the semiparametric harmonic regression model 
but the phase \eqn{\phi} is unknown, then the estimations of 
the function \eqn{g(\cdot)} and \eqn{\phi} is conducted in a iterative way.
The general optimizer \code{\link[stats]{optim}} is used to minimize the
UBRE of the fitted with respect to \eqn{\phi}.
}
\examples{
# Simulate the series
set.seed(193)
g <- function(x){-2*exp(-x^3)}
n <- 50
t <- 1:n
lambda0 <- 0.123
phi0 <- 0.2*2*pi
u <- cos(2*pi*lambda0*t+phi0)
e <- rnorm(n)
y <- g(u)+e

# Plot ther series
par(mfrow=c(1,2))
plot(t,g(u),type="b")
plot(t,y,type="b")
par(mfrow=c(1,1))

# Pre-estimate the frequency and the phase
lambdaRange <- seq(1/n,0.5-1/n,length.out = 5*n)
paraRankLS <- GetFitRankLS(y,t=1:n,lambdaRange)
paraRankLS    

# Fit the semi-harmonic regression
fit1 <- semihregScam(y,t,lambda0=paraRankLS[2],bs="mpi",family=gaussian())
names(fit1)
fit1$esty

# Prediction for the continuous time.
newt <- seq(1,n,0.05) 
head( predy <- predict(fit1,newt)) 

plot(newt,g(cos(2*pi*lambda0*newt+phi0)),
     type="b",pch=19,main="Semiparametric with scam",
     ylim=c(min(predy[,2])-1,max(predy[,2])+0.5) )
lines(newt,predy[,2],col="red",type="l")


}
\references{
Pya, N., & Wood, S. N. (2015). 
 Shape constrained additive models. Statistics and Computing, 25(3), 543-559.
}
\seealso{
\code{\link[scam]{scam}}.
}
\author{
Yuanhao Lai
}
\keyword{ts}
