% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shreg.R
\name{shreg}
\alias{shreg}
\title{Simulate harmonic regression models}
\usage{
shreg(n, f = 0, model = c("Gaussian", "ContaminatedNormal", "t5",
  "Laplace", "cubic", "AR1"), hpar = list(A = NA, B = NA, R = NA, zeta =
  NA, snr = NA), mu = 0, sig = 1, phi = 0.5, c = 3,
  g = function(x) {     x })
}
\arguments{
\item{n}{Length of series.}

\item{f}{Frequency.}

\item{model}{The model used for generating the error term. See details.}

\item{hpar}{structural parameters that specify A and B. See details.}

\item{mu}{constant term}

\item{sig}{The standard error of the series.}

\item{phi}{Only used if AR1 error distribution is selected.}

\item{c}{Only used if Contaminated Normal error distribution is selected.}

\item{g}{A function used to transform the harmonic component. See details.}
}
\value{
Vector of length n, simulated harmonic series.
}
\description{
Simulates a harmonic regression or a semiparametric harmonic regression.  
Possible types of models are normal, t(5), Laplace, cubic and AR1.
}
\details{
{ Generate a harmonic series y with length n, 
where 
\deqn{y_t = mu+g( A*cos(2*pi*f*t)+B*sin(2*pi*f*t) )+sig*e_t,\ t=1,...,n,}
and \eqn{e_t} comes from one of the following specified distributions 
with mean 0 and standard error 1 that are specified 
in the argument \code{model}:

\code{Gaussian}: A standard normal distribution (i.i.d.).

\code{ContaminatedNormal}: Generates a random sample from a normal mixture 
\eqn{p*N(0,c^2)+(1-p)*N(0,1)} where \eqn{p=1/(1+c^2))}.

\code{t5}: A t distribution with 5 degrees of freedom 
(i.i.d., standardized to mean 0 and variance 1).

\code{Laplace}: A Laplace (double exponential) distribution
(i.i.d., standardized to mean 0 and variance 1).

\code{cubic}: A standard normal distribution for e, 
but the cube transform is used on the time series.

\code{AR1}: An AR(1) series with autocorrelation paramater phi
(standardized to mean 0 and variance 1).

The argument \code{hpar} is a list that specifies the sinusoid. If components
\code{A} and \code{B} are defined, these are used. Otherwise if components
\code{R} and \code{zeta} are defined, these specify the amplitude and phase.
Finally, \code{snr} may be used to generate a model with a specified 
signal-to-noise ratio and random phase.
}
}
\examples{
#Simulate the harmonic regression model with standard Gaussian error terms
z <- shreg(10, f=2.5/10, hpar=list(snr=10))
plot(1:10,z,type="b")

#Simulate the AR(1) errors
z <- shreg(50, f=2.5/10, hpar=list(snr=2),model="AR1", phi=0.5, mu=100)
plot(z, type="o")
ans <- hreg(z)
plot(ans)
plot(ans, type="cp")

#Contaminated normal with large outlier
set.seed(77233)
z <- shreg(100, f=4/10, hpar=list(snr=10), model="ContaminatedNormal", c=25)
(out <- hreg(z))
plot(out)
(out <- hreg(z, method="MM",K=100))
plot(out)

}
\seealso{
\code{\link{hreg}}, \code{\link{ptestReg}}
}
\author{
A.I. McLeod and Yuanhao Lai
}
\keyword{ts}
